cleanup test; added new test class for certificate
