+++ src/test/java/de/braintags/vertx/util/TestAllUtil.java
@@ -21,6 +21,7 @@ import de.braintags.vertx.util.codec.AbstractPojoCodecTest;
 import de.braintags.vertx.util.file.FileSystemUtilTest;
 import de.braintags.vertx.util.lock.AsyncReadWriteLockTest;
 import de.braintags.vertx.util.security.CRUDPermissionMapTest;
+import de.braintags.vertx.util.security.CertificateTest;
 import de.braintags.vertx.util.security.crypt.impl.StandardEncoderTest;
 import de.braintags.vertx.util.tree.TreeTest;
 
@@ -35,7 +36,7 @@ import de.braintags.vertx.util.tree.TreeTest;
     ResultObjectTest.class, ErrorObjectTest.class, ExceptionUtilTest.class, AbstractCollectionAsyncTest.class,
     CollectionAsyncTest.class, FileSystemUtilTest.class, GeoLoationUtilTest.class, AbstractPojoCodecTest.class,
     AssertTest.class, CRUDPermissionMapTest.class, StandardEncoderTest.class, VertxAsyncTest.class, TreeTest.class,
-    TestResourceUtil.class, EqUtilTest.class })
+    TestResourceUtil.class, EqUtilTest.class, CertificateTest.class })
 public class TestAllUtil {
 
 }
