+++ src/test/java/de/braintags/vertx/util/security/CertificateTest.java
@@ -1,0 +1,47 @@
+/*-
+ * #%L
+ * Vert.x utilities from Braintags
+ * %%
+ * Copyright (C) 2017 Braintags GmbH
+ * %%
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ * #L%
+ */
+package de.braintags.vertx.util.security;
+
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.SecureRandom;
+import java.security.Security;
+import java.security.cert.X509Certificate;
+import java.util.Date;
+
+import org.junit.Test;
+
+/**
+ * @author Michael Remme
+ */
+public class CertificateTest {
+  private static final io.vertx.core.logging.Logger LOGGER = io.vertx.core.logging.LoggerFactory
+      .getLogger(CertificateTest.class);
+
+  @Test
+  public void certificateTest() throws Exception {
+    Security.addProvider(new org.bouncycastle.jce.provider.BouncyCastleProvider());
+
+    KeyPair pair = generateRSAKeyPair();
+    X509Certificate cert = CertificateHelper.generateV3Certificate(pair);
+    cert.checkValidity(new Date());
+    cert.verify(cert.getPublicKey());
+  }
+
+  public static KeyPair generateRSAKeyPair() throws Exception {
+    KeyPairGenerator kpGen = KeyPairGenerator.getInstance("RSA", "BC");
+    kpGen.initialize(1024, new SecureRandom());
+    return kpGen.generateKeyPair();
+  }
+
+}
